Rewrite detfm in rust
Rationale:
---------
The main reason behind this rewrite was for fun.
Using C++ is sometimes a pain: it's a fight against the build system and
the differences between GCC and Microsoft's compiler.

Rust has a great tooling for installing and managing dependencies, and
the compiler gives you terrific error messages.
It's also much easier to install a recent version of Rust, thanks to the great
rustup utility.

Drawbacks:
---------
This rewrite comes with a cost: the removal of mulitple features.
The `classdef` matching was completely removed as it was not really practical,
and only worked for simple classes.

The custom formatting of packet was also removed. Rust does not have support
for runtime string formatting with using an unstable feature or an overkill
full-featured templating system. There are a few crates for runtime formatting,
but none matches this project's needs.

Both features will come back at some point, most likely the formatting first.

Performances:
------------
The Rust port is 1.16 ± 0.02 times slower than its C++ equivalent.
However, C++ was using multiple threads to speed up the process, which is not
the case in Rust.
When disabling parallelization, the rust port is 3.12 ± 0.02 times faster :)